requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
pydantic>=2.0.0
redis>=5.0.0
aiosqlite>=0.19.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
